Fix duplicated request etc
