+++ src/schema/request.py
@@ -1,5 +1,5 @@
-from pydantic import BaseModel, EmailStr
+from pydantic import BaseModel, EmailStr, Field
 
 class KakaoLoginRequest(BaseModel):
-    email: EmailStr
-    name: str 
+    email: EmailStr = Field(..., description="User's email from Kakao")
+    name: str = Field(..., description="User's name from Kakao")
